Remove checkout layout
